Add ${folder} setting token
I used the [window variables](https://www.sublimetext.com/docs/3/api_reference.html#sublime.Window) to give the ability to `chdir` to top folder of a view when SublimeText is open in folder or project mode.

In the case of the Rubocop linter, `rubocop` uses the `.ruby-version` file [to set the targeted ruby version](https://rubocop.readthedocs.io/en/latest/configuration/#setting-the-target-ruby-version). Since the command is not run from the root folder of the project, and that the project file is not necessary inside the root folder, `${folder}` was needed to `chdir` in the right folder containing the `.ruby-version` file.